Remote transfer on ray test
